Update Pow(x, n).cpp

+++ Arrays/Pow(x, n).cpp
@@ -3,7 +3,7 @@ public:
     double myPow(double x, int n) {
         
         double ans=1;
-        int nn= n;
+        long long int nn= n;
         if(nn<0)
             nn=(-1)*nn;
         while(nn>0)
